old idiot problem happened again. Incorrectly parallel area included

+++ facs/big_query.h
@@ -11,5 +11,5 @@ extern void clean_list (Queue* head, Queue *tail);
 extern BIGCAST CHUNKer(gzFile zip,BIGCAST offset,int chunk,char *data,int type);
 extern BIGCAST CHUNKgz(gzFile zip, BIGCAST offset,int chunk,char *position,char *extra,int type);
 extern int bq_main (int argc, char **argv);
-extern int query (char* query, char* bloom_filter, float tole_rate, float sampling_rate, char* list, char* target_path);
+extern int query (char* query, char* bloom_filter, double tole_rate, double sampling_rate, char* list, char* target_path);
 #endif
